requirements.txt
torch>=2.0.0
transformers>=4.30.0
langchain>=0.0.240
sentence-transformers>=2.2.2
pypdf>=3.0.0
faiss-cpu>=1.7.4
python-dotenv>=0.19.0
accelerate>=0.21.0